+++ src/viewer.rs
@@ -1,6 +1,6 @@
 //! Module related to implementation of the [`MjViewer`] and [`MjViewerCpp`].
 
-use glfw::{Action, Context, Glfw, GlfwReceiver, Key, Modifiers, MouseButton, PWindow, WindowEvent};
+use glfw::{Action, Context, Glfw, GlfwReceiver, Key, Modifiers, MouseButton, PWindow, WindowEvent, WindowMode};
 use bitflags::bitflags;
 
 use std::time::Instant;
@@ -26,6 +26,7 @@ const DOUBLE_CLICK_WINDOW_MS: u128 = 250;
 
 const HELP_MENU_TITLES: &str = concat!(
     "Toggle help\n",
+    "Toggle full screen\n",
     "Free camera\n",
     "Track camera\n",
     "Camera orbit\n",
@@ -35,11 +36,15 @@ const HELP_MENU_TITLES: &str = concat!(
     "Object select\n",
     "Selection rotate\n",
     "Selection translate\n",
-    "Exit"
+    "Exit\n",
+    "Reset simulation\n",
+    "Cycle cameras\n",
+    "Visualization toggles",
 );
 
 const HELP_MENU_VALUES: &str = concat!(
     "F1\n",
+    "F5\n",
     "Escape\n",
     "Control + Alt + double-left click\n",
     "Left drag\n",
@@ -49,7 +54,10 @@ const HELP_MENU_VALUES: &str = concat!(
     "Double-left click\n",
     "Control + [Shift] + drag\n",
     "Control + Alt + [Shift] + drag\n",
-    "Control + Q"
+    "Control + Q\n",
+    "Backspace\n",
+    "[ ]\n",
+    "See MjViewer docs"
 );
 
 #[derive(Debug)]
@@ -80,19 +88,20 @@ impl Error for MjViewerError {
 /// the viewer doesn't store a mutable reference to the [`MjData`] struct, but it instead
 /// accepts it as a parameter at its methods.
 /// 
-/// Currently supported (to be expanded in the future):
-/// - Visualization of the 3D scene,
-/// - Close via Ctrl + Q or by closing the window,
-/// - Body tracking via Ctrl + Alt + double left click,
-/// - Camera look at object via Alt + double left click,
-/// - Escape from tracked camera via Esc.
-/// - Perturbations:
-///     - Select the object of interested by double clicking on it,
-///     - Hold down Control and start dragging for **rotational** perturbations,
-///     - Hold down Control+Alt and start dragging for **translational** perturbations,
-///     - To move in the XY plane instead of the default XZ plane, hold Shift.
-/// 
 /// The [`MjViewer::sync`] method must be called to sync the state of [`MjViewer`] and [`MjData`].
+/// 
+/// # Shortcuts
+/// Main keyboard and mouse shortcuts can be viewed by pressing ``F1``.
+/// Additionally, some visualization toggles are included, but not displayed
+/// in the ``F1`` help menu:
+/// - C: camera,
+/// - U: actuator,
+/// - J: joint,
+/// - M: center of mass,
+/// - H: convex hull,
+/// - Z: light,
+/// - T: transparent,
+/// - I: inertia.
 /// 
 /// # Safety
 /// Due to the nature of OpenGL, this should only be run in the **main thread**.
@@ -106,6 +115,7 @@ pub struct MjViewer<'m> {
     /* Other MuJoCo related */
     model: &'m MjModel,
     pert: MjvPerturb,
+    opt: MjvOption,
 
     /* Internal state */
     last_x: f64,
@@ -149,21 +159,17 @@ impl<'m> MjViewer<'m> {
         window.set_all_polling(true);
         glfw.set_swap_interval(glfw::SwapInterval::None);
 
-        let scene = MjvScene::new(model, model.ffi().ngeom as usize + scene_max_geom + EXTRA_SCENE_GEOM_SPACE);
-        let user_scene = MjvScene::new(model, scene_max_geom);
-        let context= MjrContext::new(model);
-        let camera = MjvCamera::new_free(model);
-        let pert = MjvPerturb::default();
         Ok(Self {
-            scene,
-            context,
-            camera,
-            model,
-            pert,
             glfw,
             window,
             events,
-            user_scene,
+            model,
+            scene: MjvScene::new(model, model.ffi().ngeom as usize + scene_max_geom + EXTRA_SCENE_GEOM_SPACE),
+            context: MjrContext::new(model),
+            camera: MjvCamera::new_free(model),
+            pert: MjvPerturb::default(),
+            opt: MjvOption::default(),
+            user_scene: MjvScene::new(model, scene_max_geom),
             last_x: 0.0,
             last_y: 0.0,
             status_flags: ViewerStatusBits::HELP_MENU,
@@ -232,8 +238,7 @@ impl<'m> MjViewer<'m> {
         self.update_rectangles(self.window.get_framebuffer_size());
 
         /* Update the scene from the MjData state */
-        let opt = MjvOption::default();
-        self.scene.update(data, &opt, &self.pert, &mut self.camera);
+        self.scene.update(data, &self.opt, &self.pert, &mut self.camera);
 
         /* Draw user scene geoms */
         sync_geoms(&self.user_scene, &mut self.scene)
@@ -248,7 +253,6 @@ impl<'m> MjViewer<'m> {
         rectangle.width = rectangle.width - rectangle.width / 4;
 
         /* Overlay section */
-
         if self.status_flags.contains(ViewerStatusBits::HELP_MENU) {  // Help
             self.context.overlay(
                 MjtFont::mjFONT_NORMAL, MjtGridPos::mjGRID_TOPLEFT,
@@ -275,32 +279,123 @@ impl<'m> MjViewer<'m> {
         self.glfw.poll_events();
         while let Some((_, event)) = self.events.receive() {
             match event {
+                // Set the viewer's state to pending exit.
                 WindowEvent::Key(Key::Q, _, Action::Press, Modifiers::Control) => {
                     self.window.set_should_close(true);
                     break;  // no use in polling other events
                 },
+                // Free the camera from tracking.
                 WindowEvent::Key(Key::Escape, _, Action::Press, _) => {
                     self.camera.free();
                 },
-
+                // Display the help menu.
                 WindowEvent::Key(Key::F1, _, Action::Press, _) => {
                     self.status_flags.toggle(ViewerStatusBits::HELP_MENU);
                 }
-
+                // Full screen
+                WindowEvent::Key(Key::F5, _, Action::Press, _) => {
+                    self.toggle_full_screen();
+                }
+                // Reset the simulation (the data).
+                WindowEvent::Key(Key::Backspace, _, Action::Press, _) => {
+                    data.reset();
+                    data.forward();
+                }
+                // Cycle to the next camera
+                WindowEvent::Key(Key::RightBracket, _, Action::Press, _) => {
+                    self.cycle_camera(1);
+                }
+                // Cycle to the previous camera
+                WindowEvent::Key(Key::LeftBracket, _, Action::Press, _) => {
+                    self.cycle_camera(-1);
+                }
+                // Toggles camera visualization.
+                WindowEvent::Key(Key::C, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_CAMERA);
+                }
+                WindowEvent::Key(Key::U, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_ACTUATOR);
+                }
+                WindowEvent::Key(Key::J, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_JOINT);
+                }
+                WindowEvent::Key(Key::M, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_COM);
+                }
+                WindowEvent::Key(Key::H, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_CONVEXHULL);
+                }
+                WindowEvent::Key(Key::Z, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_LIGHT);
+                }
+                WindowEvent::Key(Key::T, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_TRANSPARENT);
+                }
+                WindowEvent::Key(Key::I, _, Action::Press, _) => {
+                    self.toggle_opt_flag(MjtVisFlag::mjVIS_INERTIA);
+                }
+                // Zoom in/out
                 WindowEvent::Scroll(_, change) => {
                     self.process_scroll(change);
                 }
+                // Track cursor, possibly applying perturbations.
                 WindowEvent::CursorPos(x, y) => {
                     self.process_cursor_pos(x, y, data);
                 },
 
-                // Match left button presses
+                // Process left clicks, for selection ob bodies.
                 WindowEvent::MouseButton(MouseButton::Left, action, modifiers) => {
                     self.process_left_click(data, &action, &modifiers);
                 }
                 _ => {}  // ignore other events
             }
         }
+    }
+
+    /// Toggles visualization options.
+    fn toggle_opt_flag(&mut self, flag: MjtVisFlag) {
+        let index = flag as usize;
+        self.opt.flags[index] = !self.opt.flags[index];
+    }
+
+    /// Cycle MJCF defined cameras.
+    fn cycle_camera(&mut self, direction: i32) {
+        let n_cam = self.model.ffi().ncam;
+        if n_cam == 0 {  // No cameras, ignore.
+            return;
+        }
+
+        self.camera.fix((self.camera.fixedcamid + direction).rem_euclid(n_cam) as u32);
+    }
+
+    /// Toggles full screen mode.
+    fn toggle_full_screen(&mut self) {
+        self.glfw.with_primary_monitor(|_, m| {
+            if let Some(monitor) = m {
+                /* Obtain the new window mode outside due to lifetimes requirements */
+                let (
+                    new_mode,
+                    w, h,
+                    x, y
+                ) = self.window.with_window_mode(|mode| {
+                    // Try to obtain the monitor size, so we can center the window after.
+                    // Assume 720p (16:9) if the size can't be read.
+                    let (w, h) = monitor.get_video_mode()
+                        .map_or((1280, 720), |x| (x.width, x.height));
+                    if let WindowMode::FullScreen(_) = mode {
+                        (WindowMode::Windowed, 1280, 720, (w - 1280) / 2, (h - 720) / 2)
+                    }
+                    else {
+                        (WindowMode::FullScreen(monitor), w, h, 0, 0)
+                    }
+                });
+                self.window.set_monitor(
+                    new_mode,
+                    x as i32, y as i32,
+                    w, h, None
+                );
+            }
+        });
     }
 
     /// Processes scrolling events.
@@ -376,7 +471,7 @@ impl<'m> MjViewer<'m> {
                     /* Obtain the selection */ 
                     let rect: &mjrRect_ = &self.rect_view;
                     let (body_id, _, flex_id, skin_id, xyz) = self.scene.find_selection(
-                        data, &MjvOption::default(),
+                        data, &self.opt,
                         rect.width as MjtNum / rect.height as MjtNum,
                         (x - rect.left as MjtNum) / rect.width as MjtNum,
                         (y - rect.bottom as MjtNum) / rect.height as MjtNum