+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-The changelog is available in TODO.
+The changelog is available at [main documentation](https://mujoco-rs.readthedocs.io/en/latest/changelog.html).