+++ docs/guide/source/changelog.rst
@@ -12,6 +12,34 @@ Versioning
 =================
 This project uses `semantic versioning <https://semver.org/>`_.
 
+Unreleased
+================================
+- :docs-rs:`~mujoco_rs::viewer::<struct>MjViewer`:
+
+  - Added events for keys:
+
+    - ``Backspace``: resets the simulation,
+    - ``F5``: toggles full screen mode,
+    - ``[`` and ``]``: cycles cameras,
+    - Visualization toggles:
+
+      - ``C``: camera,
+      - ``U``: actuator,
+      - ``J``: joint,
+      - ``M``: center of mass,
+      - ``H``: convex hull,
+      - ``Z``: light,
+      - ``T``: transparent,
+      - ``I``: inertia.
+
+
+- :docs-rs:`~mujoco_rs::wrappers::mj_visualization::<type>MjvCamera`:
+
+  - Added methods:
+
+    - :docs-rs:`~mujoco_rs::wrappers::mj_visualization::<type>MjvCamera::fix`:
+      changes the camera struct parameters to display a fixed camera.
+
 1.3.0 (MuJoCo 3.3.5)
 ================================
 - Added a module for offscreen scene rendering: :docs-rs:`mujoco_rs::renderer`: